+++ src/Modules/WorkGroup/WorkGroupTransactions.php
@@ -2,7 +2,6 @@
 
 namespace Foodsharing\Modules\WorkGroup;
 
-use Foodsharing\Lib\Db\Mem;
 use Foodsharing\Modules\Bell\BellGateway;
 use Foodsharing\Modules\Bell\DTO\Bell;
 use Foodsharing\Modules\Core\DBConstants\Bell\BellType;
@@ -16,7 +15,6 @@ use Symfony\Contracts\Translation\TranslatorInterface;
 class WorkGroupTransactions
 {
     public function __construct(
-        private readonly Mem $mem,
         private readonly WorkGroupGateway $workGroupGateway,
         private readonly ForumFollowerGateway $forumFollowerGateway,
         private readonly UploadsGateway $uploadsGateway,
@@ -35,24 +33,6 @@ class WorkGroupTransactions
     {
         $this->forumFollowerGateway->deleteForumSubscription($groupId, $memberId);
         $this->workGroupGateway->removeFromGroup($groupId, $memberId);
-    }
-
-    /**
-     * Checks if an user is administrator for a working group.
-     *
-     * INFO: The reference information is cached on redis and the cache is updated by @see MaintenanceControl.
-     *
-     * @param int $userId UserId to check
-     *
-     * @return bool True is administrator, False no information present
-     */
-    public function isAdminForAWorkGroup(int $userId): bool
-    {
-        if ($allGroupAdmins = $this->mem->get('all_global_group_admins')) {
-            return in_array($userId, unserialize($allGroupAdmins));
-        }
-
-        return false;
     }
 
     public function sendMailToGroup(string $groupName, string $message, string $username, int $userId, array $recipients, string $userMail): void